Rename solidart_hooks example pages
